+++ src/app/(authorized)/profile/[id]/ProfileServer/ProfileServer.module.scss
