+++ src/app/(authorized)/profile/[id]/ProfileServer/ProfileServer.tsx
@@ -1,14 +1,14 @@
 import { Suspense } from 'react';
-import { Profile } from './Profile';
-import { Profile2 } from './Profile2';
-import { actions } from './actions';
-import { ApiResponsePosts, UserProfile } from './types';
+import { Profile } from '../Profile';
+import { Profile2 } from '../Profile2';
+import { actions } from '../actions';
+import { ApiResponsePosts, UserProfile } from '../types';
 import { headers } from 'next/headers';
-import { ProfileInfo2 } from './ProfileInfo/ProfileInfo2';
-import { ProfileInfo } from './ProfileInfo/ProfileInfo';
+import { ProfileInfo2 } from '../Skeleton/ProfileInfo/ProfileInfo';
+import { ProfileInfo } from '../ProfileInfo/ProfileInfo';
 import s from './ProfileServer.module.scss';
-import { Posts } from './Posts/Posts';
-import { LoadMore } from './Posts/load-more';
+import { Posts } from '../Posts/Posts';
+import { LoadMore } from '../Posts/load-more';
 import { findMinId } from '@/utils/findMinId';
 type Props = {
   id: number;
@@ -18,7 +18,7 @@ const ProfileServer = async ({ id, myProfile }: Props) => {
   const headersList = headers();
   const accessToken = headersList.get('accessToken');
   const userdata: UserProfile = await actions.getProfile(accessToken, id);
-  await new Promise((resolve) => setTimeout(resolve, 3000));
+  await new Promise((resolve) => setTimeout(resolve, 30000));
   const postsData: ApiResponsePosts = await actions.getPosts(id, 0);
   let minId = findMinId(postsData.items);
   return (
