+++ src/app/(authorized)/profile/[id]/Skeleton/Posts/Posts.module.scss
@@ -1,0 +1,92 @@
+@use '@/app/mixins.scss' as s;
+@keyframes placeHolderShimmer {
+  0% {
+    background-position: -800px 0;
+  }
+  100% {
+    background-position: 800px 0;
+  }
+}
+
+.imageContainer {
+  animation-duration: 2s;
+  animation-fill-mode: forwards;
+  animation-iteration-count: infinite;
+  animation-name: placeHolderShimmer;
+  animation-timing-function: linear;
+  background-color: #222;
+  background: linear-gradient(to right, #222 8%, #2d2d2d 18%, #222 33%);
+  background-size: 800px 104px;
+  aspect-ratio: 1/1;
+  position: relative;
+  border-radius: 10px;
+}
+.posts {
+  display: flex;
+  justify-content: flex-start;
+  flex-wrap: wrap;
+
+  width: 100%;
+
+  gap: 12px;
+  padding: 0px 25px;
+  margin-top: 21px;
+}
+@include s.devices(tablet-xl) {
+  .posts {
+    height: 107px;
+    padding: 0 24px;
+  }
+}
+@include s.devices(tablet) {
+  .posts {
+    gap: 4px;
+  }
+}
+
+@include s.devices(mobile) {
+  .posts {
+    padding: 0 15px;
+    gap: 2px;
+  }
+}
+
+.imageContainer {
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  max-width: 23%;
+  width: 100%;
+  aspect-ratio: 1/1;
+}
+
+.post {
+  width: 100%;
+  height: 100%;
+}
+
+.container {
+  margin: 0 auto;
+
+  &__text {
+    font-weight: 600;
+    font-size: 24px;
+  }
+}
+
+@include s.devices(tablet-xl) {
+  .imageContainer {
+    max-width: 32.2%;
+    width: 100%;
+  }
+
+  .imageContainer img {
+    max-width: 100%;
+  }
+}
+
+@include s.devices(tablet) {
+  .imageContainer {
+    max-width: 32.6%;
+  }
+}
