+++ src/app/(authorized)/profile/[id]/Skeleton/ProfileInfo/ProfileInfo.module.scss
@@ -31,7 +31,6 @@
 }
 .profile {
   display: flex;
-  position: absolute;
   .blockUser {
     .name {
       margin: 44px 0px 0px 0px;
