+++ src/app/(authorized)/profile/[id]/page.tsx
@@ -1,7 +1,8 @@
 import { Suspense } from 'react';
 import { headers } from 'next/headers';
-import { ProfileInfo2 } from './ProfileInfo/ProfileInfo2';
-import ProfileServer from './ProfileServer';
+import { ProfileInfo2 } from './Skeleton/ProfileInfo/ProfileInfo';
+import ProfileServer from './ProfileServer/ProfileServer';
+import { Posts } from './Skeleton/Posts/Posts';
 
 const Page = async ({ params }: { params: { id: string } }) => {
   const headersList = headers();
@@ -12,7 +13,12 @@ const Page = async ({ params }: { params: { id: string } }) => {
   return (
     <>
       <Suspense
-        fallback={<ProfileInfo2 myProfile={myId === id ? true : false} />}
+        fallback={
+          <>
+            <ProfileInfo2 myProfile={myId === id ? true : false} />
+            <Posts />
+          </>
+        }
       >
         <ProfileServer id={id} myProfile={myId === id ? true : false} />
       </Suspense>
