--- src/components/Skeleton/Profile/ProfileInfo/ProfileInfo.module.scss
@@ -1,91 +1,0 @@
-@use '@/app/mixins.scss' as s;
-
-@keyframes placeHolderShimmer {
-  0% {
-    background-position: -800px 0;
-  }
-  100% {
-    background-position: 800px 0;
-  }
-}
-
-.name,
-.img,
-.indicators div p:nth-child(2),
-.indicators div p:nth-child(1),
-.settings,
-.text {
-  animation-duration: 2s;
-  animation-fill-mode: forwards;
-  animation-iteration-count: infinite;
-  animation-name: placeHolderShimmer;
-  animation-timing-function: linear;
-  background-color: #222;
-  background: linear-gradient(to right, #222 8%, #2d2d2d 18%, #222 33%);
-  background-size: 800px 104px;
-  height: 70px;
-  position: relative;
-}
-.avatar {
-  display: flex;
-}
-.name {
-  width: 160px;
-  height: 20px;
-  background-color: #222;
-  margin: 43px 0px 0px 36px;
-  border-radius: 13px;
-}
-.img {
-  width: 204px;
-  height: 204px;
-  background-color: #222;
-  border-radius: 50%;
-  margin: 36px 0px 0px 25px;
-}
-
-.indicators {
-  display: flex;
-  gap: 94px;
-  margin: -150px 0px 0px 264px;
-
-  div p:first-child {
-    width: 15px;
-    height: 17px;
-    background-color: #222;
-    border-radius: 6px;
-  }
-  div p:nth-child(2) {
-    width: 76px;
-    height: 17px;
-    background-color: #222;
-    border-radius: 6px;
-    margin: 11px 0;
-  }
-}
-.btn {
-  display: flex;
-  justify-content: flex-end;
-  .settings {
-    width: 167px;
-    height: 35px;
-    background-color: #222;
-    margin: -110px 64px 0px 0px;
-    border-radius: 12px;
-  }
-}
-.descriptions {
-  margin: 18px 55px 0px 259px;
-  .text {
-    width: 97%;
-    height: 10px;
-    background-color: #222;
-    margin: 5px 5px;
-    border-radius: 12px;
-  }
-}
-@include s.devices(tablet-xl) {
-  .settings {
-    display: none;
-  }
-}
